LC-maximum-total-importance-of-roads [Runtime: 21 ms; Memory: 68.3 MB]

+++ heap-priority-queue/maximum-total-importance-of-roads.java
@@ -1,9 +1,9 @@
 class Solution {
     public class Node {
         int v; long e;
-        Node(int v, int e) {
+        Node(int v, long e) {
             this.v = v;
-            this.e = (long)e;
+            this.e = e;
         }
     }
     public long maximumImportance(int n, int[][] roads) {
